Updates contributing guidelines again
We discovered that access restrictions prevents people from assigning
issues to themselves. Instead of taking on the burden of managing access
for each contributor, it makes sense to just ask them to comment on the
issue saying they are working on it. It also changes the issue URL to
remove the no assignee filter.