Fixes #325
This changes the email form field in users#edit to render as a <p> tag
instead. It uses the form-control class so it appears as it did before.
The UsersController has had its Strong Params modified to only allow
:email as a permissible field if the current_user is a casa_admin
(role).

The Strong Params should probably be done differently, ideally as a
Pundit policy, or similar, but that was not set up for this model, and
felt out of scope.