Removes edit link for previous volunteer
In the volunteer dashboard, we could see the link to edit profile for previous volunteers assigned to the case in the case contacts section. While clicking on that link brought you right back, it's better to not provide a link in the first place. This change checks if the current user has permissions to edit the contact creator and only them provides the edit link. For all other cases, it just shows the display name of the contact creator.