Enforce server-side that at least one case must be selected
There's already client-side validation to make sure a user selects at
least one case, but this should also be present server-side in case
a browser doesn't enforce it properly.

Additionally, if a user is not assigned any cases at all, there will be
no checkboxes to select, so the client-side validation may be
inadvertently bypassed.