Closes #603 - DRYs up File Importer
I extracted subclasses for both User and Case importings, and left the
basic importing logic in the base class. Adding additional imports later
on should be less painful and involve replicating one of the subclasses.

The specs themselves were updated and refactored as well. I added
additional specs to cover the additional behaviors.