Turn off auto width for case contacts dataTable
The data table autoWidth function was calculating incorrect values for
the width of the case contacts table. Similar to how we're setting up
the `volunteersTable`, turning off autoWidth for the case contacts table
fixed the width issue we were having. Our data is now llining up with
the column headings.