Fix bugs in mobile view
This commit make the size of the page header dynamic so that in
mobile it is completely legible. Spaces are added to the top and
bottom of the login page to create spacing in mobile. Additionally,
an empty list element was added to combat a known bug in bootstrap
list-groups with height calculation on small screens.