Create test for volunteer links
Created a test to check if the volunteer name is a link. Moved the test
setup code to the top level scope.

In the index file I prepended some messages with the decorate message.
For some reason in the view test it would error with message not found.
I'm not sure what the architecture of the decorator is; however it seems
strange that when live testing it is not needed. Does anybody have any
insight on this?