Use already created Org for factories
We were creating orgs in many of the factories and I think we can just
use the already created one. We were creating ~2600 orgs per test run
and with this change we are down to ~1300. Probably not a huge
difference in overall time but as things grow, small changes like this
will add up.

I also added some FactoryBot logging in `spec/support/factory_bot.rb`.
It's pretty noisy so it's commented out but will be useful when
debugging slow tests in the future.