New workflow to unassign inactive issues.
The workflow will run once per hour and will post a warning message in
issues that are older than 5 days telling the user that the issue will
be unassigned after 7 days. After 7 days it will be unassigned. The
algorithm to determine the time of inactivity will account for comments
made but not events.