Link to CasaOrg logo with Rails path helper
As described in the Rails guide `Active Storage Overview`:
https://guides.rubyonrails.org/active_storage_overview.html#linking-to-files

This links correctly to Active Storage blobs for all Active Storage
services. The previous approach only worked for
ActiveStorage::Service::DiskService, which broke in production since
we're relying on ActiveStorage::Service::AzureStorageService.

The seeds are also now creating an Active Storage blob for the logo of
CasaOrg records. It's easier to test org_logo with this change.

Fixes #1883