Subscribe to ActionMailer events to create emails
Because we are using both Devise and custom application mailers to send
emails, simply extending the `mail` method in `ApplicationMailer` was
not sufficient to save every sent email.  None of the Devise emails were
being saved in the db.

This solution subscribes to the `process.action_mailer` event and creates
a `SentEmail` in the database once that event has fired.

This should work so long as we do not use `.deliver_later` for any of the
emails that we are logging. There is another event called
`deliver.action_mailer`, but that one only gives you the email content
itself. Although it would be possible to determine the user_id, casa_org,
mailer_type and category from the email content, it would be much more
difficult.

It is worth noting that ActionMailer interceptors and observers both also
only give you access to the email message content, so saving the SentEmail
through those would result in the same difficulties as using the
`deliver.action_mailer` event.