Fix failing system case_contacts test
In `spec/system/case_contacts/index_spec.rb` the tests were updated to
freeze time toJanuary 2, 2021.
In the previous version of the test, the `expects` were using
`I18n.l(case_contacts[0].occurred_at.to_date, format: :long)` which
caused single digit dates to contain a leading zero and thus fail the
test on the first 9 days of the month.  Freezing time to January 2, 2021
and hard-coding the expected dates ensures that this test won't fail
itermittently in the future.