Update lint hook to exit on unfixable errors
Previously, if there were linting errors that could not be fixed, the
linter would tell you, but the hook would continue and allow the git
action to go forward.  This change makes it so that if the linter cannot
auto-fix the issues, the hook terminates after the linter's error message.