Fix email reminder to volunteers w/o supervisors
If a volunteer does not have a supervisor assigned yet, and someone
sends them a reminder email and selects the "Send CC to Supervisor and
Admin" checkbox an undefined method error is thrown. This happens
because the VolunteerMailer is trying to access the `email` of the
supervisor which is `nil`.

This fix ensures that we only send the email with CC if a supervisor is
assigned to the volunteer.

Resolves #2590

Co-authored-by: Amy Gurski <amygurski@gmail.com>