Refactored Volunteer reminder form to partial
The functionality for reminding a volunteer is located on two different
forms: one for editing a volunteer and another for viewing a CASA case.
This pulls that form into a single partial so the layout and wording on
the form is consistent. This will also help with a future plan to use
this same reminder button form in a new location.

Co-authored-by: Amy Gurski <amygurski@gmail.com>