Add new slug fields to Casa orgs and cases
These new slug fields are going to be used to support friendly urls when
accessing CASA Cases. It's important to note that CASA Case Numbers are
not globally unique, so we created slugs for both the case and it's
parent organization.

Next steps: get routes working and start using these slugs

Co-authored-by: Johncarlo Cerna <johncarlocerna@gmail.com>