Adds Request Spec verifying Judge Name in DOCX for past court dates
Closes #2517

I added two new request specs to past_court_date_spec, similar to the
way they were done in the regular case report request spec. The DOCX
format is really hard to read and most of it is superfluous XML so I
wrote an additional helper method that strips out all the XML and
reveals just the underlying content. This spec was only needing to
verify that a specific content string was present, so it was enough.

That new helper is added alongside the existing (awesome!) helper for
this need.

The request specs themselves test both the affirmative and negative
cases of having a judge or not, since the judge's presence is optional.

ADDITIONAL CLEANUP:
I noticed that the PastCourtDate model had an extraneous presence
validation for the :casa_case_id field. As of Rails 5, belongs_to
implies a required presence for the association key unless you specify
optional: true. I modified the model spec that was verifying the
integrity of this association.