Remove hidden id field on new volunteer form
Remove hidden casa_org_id field on new volunteer form. Due to recent
changes the volunteer is already associated with the current organization
when it is created so this field was unused.

casa_org_id is no longer a param so update volunteer_specs accordingly.

Add a volunteer_specs test ensuring that creator's (the logged in admin's)
casa organization matches the new volunteer's organization.