Remove spring-watcher-listen gem
spring-watcher-listen doesn't allow for spring v3.0 currently but
spring v3.0 is mandatory to upgrade to rails 7. Otherwise, it's not
possible to use spring or load a rails console.

This can be added back in later if desired, but it's unlikely to be a
noticable change. Spring will use the default way of watching the filesystem
for changes without this gem. For this size app and modern computers
it's unlikely that there  will be any noticable difference in responsiveness
or cpu use on our local systems during development.

If we want to re-add it, it can be done once this issue is resolved:
https://github.com/rails/spring-watcher-listen/issues/27