Keep using imagemagick in rails 7
Rails changed their default active storage variant processor from
imagemagick to vips in rails 7. This caused the casa site to crash when manually
testing because of the vips library being unavailable. According to their docs,
it is fine to keep using imagemagick and this option is not being deprecated
so chose to keep it rather than replacing it.

https://guides.rubyonrails.org/upgrading_ruby_on_rails.html#active-storage-default-variant-processor-changed-to-vips