Fix unsafe redirects in tests using example.com
Rails 7 protects from unsafe redirects by default now. Change the tests
to use the root path as the http referrer instead of example.com so it's
allowed to fix the tests. Didn't modify the controller to allow unsafe
redirects because I'm assuming we expect the referrer when a court
date or case contact is deleted to be a URL within our site.