Fix large table overflow styling
Fix applied to the casa_cases, reimbursements, and volunteers tables.

When the window was not wide enough, some larger tables were overflowing
past their parent card div. The simple fix implemented here is to wrap
the table in a div with bootstrap's table-responsive class which makes
it automatically scroll horizontally as needed:
https://getbootstrap.com/docs/4.0/content/tables/#responsive-tables