Fix flaky reimbursement_datatable_spec.rb test
The rspec test datatables/reimbursement_datatable_spec.rb has been
failing intermittently requiring multiple CI runs to get a passing build.
The failing tests relate to ordering the reimbursement table by case number.

The test data generates 15 reimbursement request records with each
case number re-used 3 times. Grabbing the first id from the table when
sorted by the case number resulted in flakiness.

For the purposes of the test, consider it sorted if the smallest case number
is the first record when ordered ascending and the largest case number
is the first record when ordered descending. This means it could be any
one of 3 records but still proves that it's ordering by case number so
should be sufficient to meet the purpose of these "ordered by case number"
tests.