Close #3449 Replace translation with literal
When reading and modifying code it can be confusing to see a translation lookup. Instead this commit changes the inactive_case to use string literals instead. This makes modifying the view easier for developers, though comes at the cost of reduced ability to use localization features to switch language output. However, since there are only English locales, this commit represents no functional change to the application.