Address flay system specs by disabling Rack::Attack in the test environment
This commit disables Rack::Attack in the test environment except for a single spec that needs it.

The problem we noticed was that system specs occasionally fail in the CI and the resulting screenshot at the failure is a blank page that just says "Retry later". This seems similar to an issue posted at https://github.com/thoughtbot/capybara-webkit/issues/628 and the fix for that issue was to disable Rack::Attack in the test environment.

It's difficult to be certain that this will fix the specs since we aren't able to reproduce locally but this seems like a reasonable thing to try.