Fix flaky learning hour test
The learning hour name is randomly generated from faker book titles.
Sometimes, the titles have commas in them. When the csv is generated by
the LearningHoursExportCsvService, it will put quotes around the name
if they contain a comma. The tests didn't expect these quotes so they would
fail intermittently.

Refactor the tests to handle cases where there is a comma in the name.