showed the add a court date after the court dates inside the casa_case
I didn't find anywhere to add the new court date, but
this button that was being hidden. So, for now, I'm only
showing it.

I also changed a bit how the button is structured, I thought
it was a bit confusing with the big add sign inside the button,
and the text outside.