updated casa_cases/edit_spec & new_spec & show
updated the tests to expect the due date to not be present.
since they are now being displayed in the court date, instead of
the main case.

also updated the show test: I had previously deleted the test,
but we should make sure it isn't present. I think it also makes
the history of these changes clearer.

and updated the case show, since we are viewing the due date
in the court date.