added nested form fields to edit court_report_due_date
The whole problem here was, the court_report_due_date is
a column from casa_case. So when we moved it to the
court_date, it wasn't updating.

But since our tests also check for the nil value
when the court_report_due_date doesn't exist, it took
me a while to realize the problem.

Previously I had only tried to nest forms, that obviously
didn't work. So I figured out how to use the form_fields
for subforms.

For it to work, I had to add the casa_case_attributes
court_report_due_date to the court_date controller.
And also had to add accepts_nested_attributes_for
in the court_date model, with update only, because we
only add a court_report_due_date to an existing casa_case
and court_date.