coupled input and label from emancipation_category
This way we can click on input or label, but the '+' remains
a separate thing.

So when we click '+' it only expands the inputs, but when
we click either the label or the input it checks the input
and expands the inner inputs.

Issue: The css needs some fixing, the '+' is out of place.