Redirect to Case Contacts page if more than one case is selected.
The case page has a "new case contact" button that filters to the
current case. Currently, when creating a new case contact, if there are
multiple cases selected, we redirect to the first case. If a volunteer
is entering case contacts for multiple cases, they have to keep going to
the case contacts page, because the "new case contact" button on the
case page only shows one of the two cases.

Instead, if they have entered case contacts for multiple cases, we
should take them to the case contacts page, so they can continue
entering cases for multiple contacts.

Added the Thank You modal to the case contacts index page so success
messages will show up there too, and moved supporting code to a helper.
(It was independent and didn't need to be in a decorator.)