fixed data-is-open="false" on when input checked saved
When we check the input and/or it's options, it saves that state.
So when we close and open the page again, the options are not hidden.
However, we were getting a data-is-open="false" by default, even
when it wasn't. So I added a check to add the proper boolean
the first time around.

However #2, manually testing, it seems to work, but the automated
test still doesn't work