refactored to remove loop
If the last court date is the past, and the query returns them
in ascending order, then all are in the past. If the last is in
the future, then the first could be in the past, so we check for that.
So we only need to check the first and last court_dates, not all.