fix: update CasaCasesController Index authorization
Not sure why we previously used `authorize CasaCase` as it was passing in a
CasaCase:Class which was not helpful at all for authorization purposes. It
worked because the `application_policy#index?` method checked for
`is_admin?`. Because we are wanting to scope the entire controller by
organization, we should pass in `current_organization` and specify that
we want to use the `CasaCasePolicy` object. This way, the `index?`
authorization from the correct policy is used.

With this change, the entire controller authorizes using an instance of
some Rails model.

Co-authored-by: Gustavo Valenzuela <gustavovalenzuela1180@gmail.com>