fix: Add behaviour to `same_org?` policy
This commit:
* Reverts the `CasaCasesController#index` action authorization back to the
  original form.
* Updates `same_org?` in `ApplicationPolicy` to handle a collection
  authorization request by adding a `Class` pattern match case.
* Updates `case_case_policy_spec.rb` to conceed that we only check a
  user's role when doing collection related policies (like `index`)
* Restores some comments in `CasaCasePolicy` for reference purposes

Because
The previous change we made to `ApplicationPolicy` expects too much
change to areas of the application we are currently not working on which
can be a problem.

The pattern of using `authorize Model` is suggested by Pundit for
collection actions/policies, and we should stick with letting the
application authorize using this pattern, rather than change the policy
to always expect an authorization subject/target especially because there
are other places in the app that use this pattern to authorize.

As a result, when checking against a collection we need to handle cases
where the record in question might not be a model instance. In all
cases, a `nil` record returns a not authorized, as it seems safer to
automatically deny a `nil` record than to allow.

For a `Class` type record, we are then authorizing a collection policy.
In this case, the `current_organization` is actually dervied from the
current user so it will always be true, because checking
`user.casa_org == current_organization` is more or less asking if the
two variables point to the same object. We just return `true` for that
reason.

Co-authored-by: Gustavo <gustavovalenzuela1180@gmail.com>