fix: change policy check for update_contact_types
Previously, we were calling Pundit explicitly and passing both
`current_user` and the `casa_case` but this was not necessary, as
Pundit's `policy` method automatically captures the current user.

We think this was done to help the view spec pass, that spec was
not set up correctly, which has been resolved.

Co-authored-by: Gustavo Valenzuela <gustavovalenzuela1180@gmail.com>