Guarantee case contacts before and after most recent past court date in seed data (#4796)
Add active record query methods to determine if there was a historic court date and return the date
of the most recent court date if so; determine if there were any case contacts before and after this court
date. Add conditional to check if there was a historic court date and create a new case contact before and
after this date if neither exists.