fix(tests): webmock test flakes
The previous behavior in terms of making API calls in the application
was a allowing all external calls in test and blacklisting certain
requests if the tests were making those calls.

This resulted in several issues:
- some tests were not blacklisting all the calls they should have been
so we were making 3rd party calls in our specs (woops)
- some tests were not returns the allow_list back to the default state
which resulted in localhost calls failing

I have changed the strategy to deny all exertanal calls (except
localhost and docker). All requests will needed to be stubbed using
WebMock.