refactor: make truncation component less generic
My initial idea was to make a generic truncation component that could be
reused in many places. This led to a more complex interface for using
the component.

At this point it is only used in one place so it makes the most sense to
tailor it to that use case.