feat: add contact notes to case contact card (#5816)
* feat: add contact notes to case contact card

Created a component to enable text truncation behavior as well as
showing the full text. This is done with CSS rather than server side.

Shows notes and answers to contact topics on a per case contact basis.

* fix: removes outdated tests

Testing for truncation is done in component specs, these tests can be
removed.

HTML should be escaped if it is submitted in notes. It is unclear to me
why we are allowing HTML to get parsed and displayed.

* refactor: make truncation component less generic

My initial idea was to make a generic truncation component that could be
reused in many places. This led to a more complex interface for using
the component.

At this point it is only used in one place so it makes the most sense to
tailor it to that use case.