fix: authorize mark_as_read in Notification controller
When we upgraded Noticed we added skipping of callbacks to the
NotificationsController. This resulted in issues with unauthed users
hitting the notifications page.

To fix this we removed the skipping of callbacks, however, one of the
callbacks being skipped was a callback to authorize the Notification.
Authorization logic had never been added so now mark as read would not
work.

This PR added authorization logic as well as request and policy specs
for the controller.