feat: use rspec retry to deflake ci (#6285)
Adds rspec-retry to compensate for some flakiness by
retrying failures.

The basic issue in common with the failures is a pattern like:

- perform action on page
- assert some database state

The issue is that the tests are often faster than the server, 
thus, the check it the action worked happens before the 
action finished. Ex:

- click submit to create case contact
- controller gets the request and starts working
-  check that count of contacts went up => fails controller not done yet
- controller finishes

Thus, we can use capybara waits to fix this. Basically we want
 to check that the front end shows the action finished 
(we see an alert) then check the database state.

- perform action on page
- check for front end flash message (capybara will wait on these)
- assert some database state